more CSS work on movie info page

+++ public/javascripts/reviews_dynamic.js
@@ -1,4 +1,4 @@
-// const db = require('../../db/models');
+
 
 window.addEventListener("load", async (event) => {
   const currentMovieId = window.location.href.split("/");
@@ -10,7 +10,7 @@ window.addEventListener("load", async (event) => {
     `/reviews/${currentMovieId[currentMovieId.length - 1]}`
   );
   const jsonReviewsData = JSON.parse(await reviewsData.json());
-//   console.log(jsonReviewsData);
+
 
   for (let el of jsonReviewsData) {
     const pTag = document.createElement("p");
@@ -25,6 +25,7 @@ window.addEventListener("load", async (event) => {
     editBtn.className = `editbtn userid-${el.userId}`;
     editBtn.innerText = "Edit";
     deleteBtn.className = `deletebtn userid-${el.userId}`;
+    deleteBtn.setAttribute('id', `deletebtn-reviewId-${el.id}`)
     deleteBtn.innerText = "Delete";
     pTag.appendChild(liTag);
     pTag.appendChild(editBtn);
@@ -40,26 +41,29 @@ window.addEventListener("load", async (event) => {
     });
   });
 
-const deleteBtns = document.querySelector('.deletebtn')
 
-for (let i = 0; i < deleteBtns.length; i++) {
-    const btn = deleteBtns[i];
-    btn.addEventListener('click', async(e) => {
-        // console.log(reviewId)
-        const res = await fetch(`/reviews/${currentMovieId[currentMovieId.length - 1]}`, {
-            method: 'DELETE'
-        })
 
-        const data = await res.json()
-        if (data.message === "Success") {
-            const container = document.querySelector(`.reviews_description`)
-            container.remove()
-        } else {
-            const pTag = document.createElement("p");
-            pTag.innerText = 'Error';
-            reviewsDiv.appendChild(pTag);
-        }
-    })
-}
+
+const reviewsDescription = document.querySelector('.reviews_description');
+  reviewsDescription.addEventListener('click', async (e) => {
+
+    const btn = e.target.id.split('-')[0];
+    if (btn === 'deletebtn') {
+      const reviewsId = e.target.id.split('-')[2]
+      const res = await fetch(`/reviews/${reviewsId}`, {
+        method: 'DELETE'
+      });
+
+      const data = await res.json()
+      if (data.message === "Success") {
+        const container = document.getElementById(`reviewId-${reviewsId}`);
+        container.remove();
+      } else {
+        const pTag = document.createElement("p");
+        pTag.innerText = 'Error';
+        reviewsDiv.appendChild(pTag);
+      }
+    }
+  });
 
 });
